backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
supabase==2.4.2
python-jose[cryptography]==3.3.0
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
openai==1.30.1
razorpay==1.4.1
httpx==0.27.0

# Optional — install separately if you want Prophet forecasting
# prophet==1.1.5
